+++ src/commands/helps/publish.rs
@@ -1,13 +1,13 @@
 // This help page was automatically generated from the mangen.py tool.
-pub const HELP: &str = r#"Release an ip through a channel.
+pub const HELP: &str = r#"Post an ip to a channel.
 
 Usage:
-    orbit launch [options]
+    orbit publish [options]
 
 Options:
-    --ready         perform a full run
+    --ready, -y     perform a full run
     --no-install    skip installing the ip
     --list          view available channels and exit
 
-Use 'orbit help launch' to read more about the command.
+Use 'orbit help publish' to read more about the command.
 "#;