+++ src/commands/manuals/publish.rs
@@ -1,24 +1,24 @@
 // This manual page was automatically generated from the mangen.py tool.
 pub const MANUAL: &str = r#"NAME
-    launch - release an ip through a channel
+    publish - post an ip to a channel
 
 SYNOPSIS
-    orbit launch [options]
+    orbit publish [options]
 
 DESCRIPTION
     Performs a series of checks on a stable version of a local ip to then release it
     through a channel.
     
-    For an ip to be launched, it must have its source field defined that directs to
+    For an ip to be published, it must have its source field defined that directs to
     a valid internet location.
     
     By default, it operates a dry run, performing all steps in the process except
     for the actual release through the channel. To fully run the command, use the
-    '--ready' flag. When the ip is launched, it will also be installed to the cache
+    '--ready' flag. When the ip is published, it will also be installed to the cache
     by default. To skip this behavior, use the '--no-install' flag.
 
 OPTIONS
-    --ready
+    --ready, -y
         Perform a full run
 
     --no-install
@@ -28,6 +28,6 @@ OPTIONS
         View available channels and exit
 
 EXAMPLES
-    orbit launch
-    orbit launch --ready
+    orbit publish
+    orbit publish --ready
 "#;